R7: slirp runs on Win64 for the first time, in a process of its own
    addresses (all must fit in 32 bits):
      a static (client_addr)   0000000000421940
      a heap block             000000000045DC90
      code (this function)     0000000000402430
      ok
    slirp_init() returned
    ARP:  reply from 52:54:00:12:35:02  ok
    DHCP: OFFER of 10.0.2.15, 590 bytes  ok
    slirp selftest: all checks passed

Nothing here touches the driver or a guest. That is the point: this is ten
thousand lines of 2004 network code that had never executed on this
architecture, so it runs where a crash costs a test program.

Three things were wrong, and only one of them was the one I went looking for.

The first-order killer was the bitfield layout, and I had not even suspected
it. struct ip and struct tcphdr declare their two nibbles as `u_int x:4, y:4`,
and mingw-w64 defaults to -mms-bitfields, which gives that pair a whole
four-byte storage unit: struct ip measures 24 bytes with ip_src at offset 16
where the wire has it at 12, struct tcphdr 24 with th_flags at 16 rather than
13. Every datagram in both directions parses as garbage, at any setting of
anything else. They are u_int8_t pairs now, which occupy one byte under the MS
and GNU layouts alike -- and struct ipasfrag already carried a comment from
someone hitting this on Linux 2.0 and fixing that one struct only.

The second was SIZEOF_CHAR_P, hardcoded to 4 in slirp_config.h with "XXX:
patch it" beside it, and nothing in the build overriding it. Both values are
broken here. At 4, caddr32_t becomes a real 8-byte pointer under LLP64 and
struct ipovly grows from 20 bytes to 32 -- two pointers plus the alignment
padding they force -- so the overlays that are supposed to land on wire format
do not. At 8 the overlays are right and the 32-bit queue-link fields inside
them go back to holding what they were designed to hold, but insque_32 and
remque_32 then store pointers through those fields and dereference the
remains. It is 8 now, with the reason written where the define is.

Which needs the third piece. Those truncated pointers are not a heap gamble
that might come out low: mingw-w64's default image base is 0x140000000, so the
IP fragment queue head -- a static -- is above 4 GB before any allocation
happens. So the daemon is linked at 0x400000 and has
IMAGE_FILE_LARGE_ADDRESS_AWARE cleared, which confines the process to 2 GB and
makes every pointer in it fit in 32 bits. ld can clear that flag itself for
i386 and the x86-64 linker rejects the option outright, so tools/pe-clear-laa.py
flips the bit, chained onto the link so it cannot be forgotten. The
confinement costs nothing: this daemon relays frames between a socket and the
monitor and maps no guest memory.

Five _Static_asserts in slirp.c now pin ip==20, tcphdr==20, ipovly==20,
tcpiphdr==40, udpiphdr==28, ipasfrag==20. Five lines that would have turned
this entire evening into a build error.

The self-test itself is -t on the slirp daemon, and it checks the two things
that need no network. ARP is answered inside slirp.c without touching the IP
path, so it is the narrowest possible "is slirp alive". DHCP is the one that
matters: a DISCOVER traverses ip_input, udp_input and bootp_input, so an OFFER
coming back proves struct ip, struct udphdr and struct udpiphdr all parse in
both directions -- and a guest with no lease resolves no names, so this is
also the first thing pacman will need. The DISCOVER is assembled from byte
offsets rather than from slirp's own structs, because those structs are the
thing under test and building the input with them would hide the errors being
hunted. On failure it prints the frames rather than a verdict.

Also verified: the addresses above are real output from the box, not an
argument that they should be low. And two claims of mine that turned out to be
wrong are recorded here so nobody re-derives them -- slirp storing winsock
SOCKET handles in int is fine, since Windows socket handles fit in 32 bits;
and slirp's pervasive assumption that long is 32 bits actually holds under
LLP64, which makes this code safer on Win64 than on 64-bit Linux.

Untested still: every line of TCP. ARP and DHCP exercise the IP and UDP
layouts; the reassembly path and the socket layer wait for a real download.

+++ src/colinux/user/slirp/ip.h
@@ -76,12 +76,24 @@ typedef u_int32_t n_long;                 /* long as received from the net */
 /*
  * Structure of an internet header, naked of options.
  */
+/*
+ * The two nibbles share one byte, and the type has to say so.
+ *
+ * Declared as u_int:4 these compile to a four-byte storage unit under
+ * mingw-w64, which defaults to -mms-bitfields: struct ip measures 24 bytes
+ * with ip_src at offset 16 where the wire has it at 12, and every datagram
+ * parses as garbage in both directions. A u_int8_t pair occupies one byte
+ * under the MS and the GNU layout alike, which is what the wire format says
+ * and what every other field in these headers already assumes. The comment
+ * in struct ipasfrag below records someone hitting this on Linux 2.0 and
+ * fixing that one struct only.
+ */
 struct ip {
 #ifdef WORDS_BIGENDIAN
-	u_int ip_v:4,			/* version */
+	u_int8_t ip_v:4,		/* version */
 		ip_hl:4;		/* header length */
 #else
-	u_int ip_hl:4,		/* header length */
+	u_int8_t ip_hl:4,		/* header length */
 		ip_v:4;			/* version */
 #endif
 	u_int8_t ip_tos;			/* type of service */
@@ -237,10 +249,10 @@ struct ipq {
  */
 struct	ipasfrag {
 #ifdef WORDS_BIGENDIAN
-	u_int	ip_v:4,
+	u_int8_t ip_v:4,
  		ip_hl:4;
 #else
-	u_int	ip_hl:4,
+	u_int8_t ip_hl:4,
 		ip_v:4;
 #endif
                                         /* BUG : u_int changed to u_int8_t.
